refactor: adjust folder organization
