+++ backend/src/data/protocols/db/user/db-load-user-by-email-and-registration.ts
